% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schextra_utils.R
\name{.schextra_bin}
\alias{.schextra_bin}
\title{Bivariate binning of single cell data into hexagon cells.}
\usage{
.schextra_bin(obj, nbins, dr, use_dims)
}
\arguments{
\item{nbins}{The number of bins partitioning the range of the first
component of the chosen dimension reduction.}

\item{use_dims}{A vector of two integers specifying the dimensions used.}

\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{dimension_reduction}{A string indicating the reduced dimension
result to calculate hexagon cell representation of.}
}
\value{
A schex-style binning list.
}
\description{
\code{make_hexbin} returns a
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object of binned hexagon cells.
}
\details{
...
}
