% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_schextra_density.R
\name{plot_schextra_density}
\alias{plot_schextra_density}
\title{Plot of density of observations from single cell data
in bivariate hexagon cells.}
\usage{
plot_schextra_density(
  obj,
  nbins = 80,
  dimension_reduction = "UMAP",
  use_dims = c(1, 2),
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{obj}{A SCUBA-supported single-cell object.}

\item{nbins}{The number of bins partitioning the range of the first
component of the chosen dimension reduction.}

\item{dimension_reduction}{A string indicating the reduced dimension
result to calculate hexagon cell representation of.}

\item{use_dims}{A vector of two integers specifying the dimensions used.}

\item{title}{A string containing the title of the plot.}

\item{xlab}{A string containing the title of the x axis.}

\item{ylab}{A string containing the title of the y axis.}
}
\value{
A \code{\link{ggplot2}{ggplot}} object.
}
\description{
Plot of density of observations from single cell data
in bivariate hexagon cells.
}
